cleanup unused library; setup mock IM format; refactor to use an IM setup; add in 'typing' message for fun
