+++ client/package.json
@@ -9,7 +9,6 @@
   "dependencies": {
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
-    "sillyname": "^0.1.0",
     "socket.io-client": "^2.0.4"
   },
   "scripts": {
